[numerics] Move MultiJac/MultiNewton
Both are now used for oneD and zeroD, so the old oneD location is no
longer appropriate.
