+++ src/numerics/MultiJac.cpp
@@ -3,7 +3,7 @@
 // This file is part of Cantera. See License.txt in the top-level directory or
 // at https://cantera.org/license.txt for license and copyright information.
 
-#include "cantera/oneD/MultiJac.h"
+#include "cantera/numerics/MultiJac.h"
 #include <ctime>
 
 namespace Cantera
